+++ credativ_pg_migrator/constants.py
@@ -17,7 +17,7 @@
 class MigratorConstants:
     @staticmethod
     def get_version():
-        return '0.10.1'
+        return '0.10.2'
 
     @staticmethod
     def get_full_name():